feat: Layout 컴포넌트 구현 및 적용
