+++ src/pages/CardRegistration/CardRegistration.tsx
@@ -2,11 +2,10 @@ import { useState } from 'react';
 import useModal from '../../hooks/useModal';
 import BottomSheet from '../../components/BottomSheet';
 import CardRegistrationForm from '../../components/CardRegistrationForm';
-import type { CardType } from '../../domain/types/card';
-import Header from '../../components/Header';
 import { ReactComponent as ChevronLeft } from '../../assets/chevron-left.svg';
-import styled from 'styled-components';
 import BankProfileList from '../../components/BankProfileList';
+import Layout from '../../components/Layout/Layout';
+import type { CardType } from '../../domain/types/card';
 
 type CardRegistrationProps = {
   setPageCardAlias: () => void;
@@ -19,11 +18,10 @@ const CardRegistration = ({ setPageCardAlias, setPageCardList, setCurrentId }: C
   const [cardType, setCardType] = useState<CardType>('우리카드');
 
   return (
-    <Styled.Wrapper>
-      <Styled.HeaderWrapper>
-        <ChevronLeft width="16px" height="16px" cursor="pointer" onClick={setPageCardList} />
-        <Header title="카드 추가" />
-      </Styled.HeaderWrapper>
+    <Layout
+      title="카드 추가"
+      leftButton={<ChevronLeft width="16px" height="16px" cursor="pointer" onClick={setPageCardList} />}
+    >
       <CardRegistrationForm
         setPageCardAlias={setPageCardAlias}
         setCurrentId={setCurrentId}
@@ -35,25 +33,8 @@ const CardRegistration = ({ setPageCardAlias, setPageCardList, setCurrentId }: C
           <BankProfileList closeModal={closeModal} setCardType={setCardType} />
         </BottomSheet>
       )}
-    </Styled.Wrapper>
+    </Layout>
   );
 };
 
 export default CardRegistration;
-
-const Wrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-  padding: 20px;
-`;
-
-const HeaderWrapper = styled.div`
-  display: flex;
-  gap: 10px;
-`;
-
-const Styled = {
-  Wrapper,
-  HeaderWrapper,
-};
